доработка структуры проекта, добавление пакета classnames, добавление компонента формы
