+++ project/src/components/app/index.ts
@@ -1,0 +1,3 @@
+import App from './app';
+
+export default App;
