+++ project/src/components/place-card-list/index.ts
@@ -1,0 +1,3 @@
+import PlaceCardList from './place-card-list';
+
+export default PlaceCardList;
