+++ project/src/components/place-card/index.ts
@@ -1,0 +1,3 @@
+import PlaceCard from './place-card';
+
+export default PlaceCard;
