+++ project/src/components/private-route/index.ts
@@ -1,0 +1,3 @@
+import PrivateRoute from './private-route';
+
+export default PrivateRoute;
