+++ project/src/components/review-form/index.ts
@@ -1,0 +1,3 @@
+import ReviewForm from './review-form';
+
+export default ReviewForm;
