+++ project/src/pages/favorites-empty/favorites-empty.tsx
