+++ project/src/pages/favorites/favorites.tsx
@@ -1,0 +1,50 @@
+import Header from '../../components/header';
+import offers from '../../mocks/offers';
+import PlaceCardList from '../../components/place-card-list';
+
+const Favorites = () => {
+  const favoriteCards = offers.slice(0,3);
+  return (
+    <div className="page">
+      <Header isLoginPage={false}/>
+
+      <main className="page__main page__main--favorites">
+        <div className="page__favorites-container container">
+          <section className="favorites">
+            <h1 className="favorites__title">Saved listing</h1>
+            <ul className="favorites__list">
+              <li className="favorites__locations-items">
+                <div className="favorites__locations locations locations--current">
+                  <div className="locations__item">
+                    <a className="locations__item-link" href="#">
+                      <span>Amsterdam</span>
+                    </a>
+                  </div>
+                </div>
+                <PlaceCardList cardList={favoriteCards} classTitle="favotites" />
+              </li>
+
+              <li className="favorites__locations-items">
+                <div className="favorites__locations locations locations--current">
+                  <div className="locations__item">
+                    <a className="locations__item-link" href="#">
+                      <span>Cologne</span>
+                    </a>
+                  </div>
+                </div>
+                <PlaceCardList cardList={favoriteCards} classTitle="favorites" />
+              </li>
+            </ul>
+          </section>
+        </div>
+      </main>
+      <footer className="footer container">
+        <a className="footer__logo-link" href="main.html">
+          <img className="footer__logo" src="../project/public/img/logo.svg" alt="6 cities logo" width="64" height="33"/>
+        </a>
+      </footer>
+    </div>
+  );
+};
+
+export default Favorites;
