+++ project/src/pages/favorites/index.ts
@@ -1,0 +1,3 @@
+import Favorites from './favorites';
+
+export default Favorites;
