+++ project/src/pages/main-empty/index.ts
@@ -1,0 +1,3 @@
+import MainEmpty from './main-empty';
+
+export default MainEmpty;
