+++ project/src/pages/main/main.tsx
@@ -68,7 +68,7 @@ const Main = ({cardList}: PropsType) => (
                 <li className="places__option" tabIndex={0}>Top rated first</li>
               </ul>
             </form>
-            <PlaceCardList cardList={cardList}/>
+            <PlaceCardList cardList={cardList} classTitle="cities"/>
           </section>
           <div className="cities__right-section">
             <section className="cities__map map"/>
