+++ project/src/pages/not-found-page/index.ts
@@ -1,0 +1,3 @@
+import NotFoundPage from './not-found-page';
+
+export default NotFoundPage;
