+++ project/src/pages/room-not-logged/index.ts
@@ -1,0 +1,3 @@
+import RoomNotLogged from './room-not-logged';
+
+export default RoomNotLogged;
