+++ project/src/pages/room/index.ts
@@ -1,0 +1,3 @@
+import Room from './room';
+
+export default Room;
